# Conflicts:
#	pom.xml
#	src/main/java/org/sanmarcux/samples/sakila/SakilaApplication.java
#	src/main/java/org/sanmarcux/samples/sakila/dao/StaffRepository.java

+++ src/main/java/org/sanmarcux/samples/sakila/business/FilmBusiness.java
@@ -17,7 +17,7 @@ public interface FilmBusiness {
 
     FilmDTO create(FilmDTO payload);
 
-    List<FilmDTO> findFilmsByActor(Short actorId);
+    List<FilmDTO> findFilmsByActor(Integer actorId);
 
-    FilmDTO get(Short filmId);
+    FilmDTO get(Integer filmId);
 }
